Bypass GIC in Pi 4
Until RISC OS supports the GIC interrupt controller in the Pi 4, we need to
set it up in bypass mode. This can only be done from secure mode, which can't
be entered once we get as far as executing the RISC OS image. This means we
have to patch the bootloader's armstubs.

To make use of this, copy the armstub8-32-gic.bin that is built by this project
into your boot partition as pi4.bin and add the following lines to the end
of your config.txt:

[pi4]
armstub=pi4.bin